Implements weighable product support and scale export
Upgrades the inventory system to support fractional stock quantities, enabling better management of weighable products.

*   Changes various quantity fields (product stock, sale items, ledger entries, reports) from integer to float64 to accurately represent partial units.
*   Introduces an `IsWeighable` flag on products to categorize items sold by weight.
*   Adds a new API endpoint `/api/products/scale-export` that generates a CSV file of weighable products, formatted for direct import into retail scale software.
*   Updates the AI agent's product creation and stock handling logic to align with the new float64 quantity type.

+++ internal/ai/agent.go
@@ -113,7 +113,7 @@ func RunAgent(userMessage string, apiKey string) (string, error) {
 				type SimpleProduct struct {
 					ID    uint    `json:"id"`
 					Name  string  `json:"name"`
-					Stock int     `json:"stock"`
+					Stock float64 `json:"stock"` // UPGRADED: Float64
 					Price float64 `json:"price"`
 				}
 				var simpleList []SimpleProduct
@@ -199,7 +199,7 @@ func executeCreateProduct(ctx context.Context, session *genai.ChatSession, funcC
 		Name:          args["name"].(string),
 		Price:         args["price"].(float64),
 		Category:      args["category"].(string),
-		StockQuantity: int(args["stock_quantity"].(float64)),
+		StockQuantity: args["stock_quantity"].(float64), // UPGRADED: Float64
 		ImageURL:      "https://via.placeholder.com/150",
 	}
 	database.DB.Create(&newProd)
